+++ .github/workflows/cpython.yaml
@@ -1,0 +1,38 @@
+name: CPython runner
+
+on:
+  release:
+    types: [ published ]
+
+jobs:
+  build:
+    strategy:
+      matrix:
+        os: [ ubuntu-latest ]
+    runs-on: ${{ matrix.os }}
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up JDK 14
+        uses: actions/setup-java@v2
+        with:
+          java-version: '14'
+          distribution: 'adopt'
+      - uses: actions/setup-python@v2
+        with:
+          python-version: '3.8.10'
+      - name: Build with Maven
+        run: mvn clean verify -B -Pcpython
+      - name: archive artifacts
+        uses: actions/upload-artifact@v2
+        with:
+          name: artifacts
+          path: |
+            transpiler/src/test/resources/org/polystat/py2eo/transpiler/simple-tests/*/genCageEO/*.eo
+            scalastyle-output.xml
+      - name: Rename jar
+        run: cp transpiler/target/transpiler-*-jar-with-dependencies.jar transpiler.jar
+      - name: archive jar
+        uses: actions/upload-artifact@v2
+        with:
+          name: package
+          path: transpiler.jar
